+++ packages/adminui-vue/components/UAutoField.vue
@@ -31,7 +31,7 @@ export default {
     },
 
     /**
-     * If defined then specified component will be used instead of default what based on attribute type.
+     * specify a component what should be used instead of default, based on attribute type.
      * For example `<u-auto-field attribute-name="bool_attr" force-cmp="el-switch" />` will create
      * `el-switch` instead of `el-checkbox` (default cmp for Boolean)
      */
@@ -140,6 +140,8 @@ export default {
 
   render (h) {
     let cmp
+    let defIsRequired = this.isRequired
+    const /** @type {UBEntityAttribute} */ATTR = this.entitySchema.attributes[this.attributeName]
     const baseAttrs = { // vue split attrs into attrs and props automatically
       ...this.$attrs,
       attributeName: this.attributeName,
@@ -148,10 +150,15 @@ export default {
       readonly: this.isDisabled || this.$attrs.readonly || this.isReadOnly,
       required: this.isRequired
     }
-    switch (this.dataType) {
+    switch (ATTR.dataType) {
       case 'Boolean':
         if (!baseAttrs.disabled && baseAttrs.readonly) {
           baseAttrs.disabled = baseAttrs.readonly // need because 'el-checkbox' and 'el-switch' doesn't have 'readonly' prop
+        }
+        if ((this.required === undefined) && ATTR.defaultValue) {
+          // hide asterisk for boolean attributes with defaultValue specified (as should be in most case)
+          // and `required` prop for UAutoField is not specified explicitly
+          defIsRequired = false
         }
         cmp = h(this.forceCmp || 'el-checkbox', {
           attrs: baseAttrs,
@@ -162,8 +169,8 @@ export default {
       case 'DateTime':
         cmp = h(this.forceCmp || 'u-date-picker', {
           attrs: {
-            type: this.dataType.toLowerCase(),
-            placeholder: this.$ut(this.dataType === 'Date' ? 'selectDate' : 'selectDateAndTime'),
+            type: ATTR.dataType.toLowerCase(),
+            placeholder: this.$ut(ATTR.dataType === 'Date' ? 'selectDate' : 'selectDateAndTime'),
             ...baseAttrs
           },
           on: this.buildListenersOnInput
@@ -172,8 +179,8 @@ export default {
       case 'Enum':
         cmp = h(this.forceCmp || 'u-select-enum', {
           attrs: {
-            eGroup: this.entitySchema.attributes[this.attributeName].enumGroup,
-            clearable: this.entitySchema.attributes[this.attributeName].allowNull,
+            eGroup: ATTR.enumGroup,
+            clearable: ATTR.allowNull,
             ...baseAttrs
           },
           on: this.buildListenersOnInput
@@ -182,7 +189,7 @@ export default {
       case 'Entity':
         cmp = h(this.forceCmp || 'u-select-entity', {
           attrs: {
-            entityName: this.associatedEntity,
+            entityName: ATTR.associatedEntity,
             ...baseAttrs
           },
           on: this.buildListenersOnInput
@@ -191,7 +198,7 @@ export default {
       case 'Many':
         cmp = h(this.forceCmp || 'u-select-many', {
           attrs: {
-            entityName: this.associatedEntity,
+            entityName: ATTR.associatedEntity,
             ...baseAttrs
           },
           on: this.buildListenersOnInput
@@ -222,7 +229,7 @@ export default {
       case 'String':
         cmp = h(this.forceCmp || 'u-input', {
           attrs: {
-            maxLength: this.entitySchema.attributes[this.attributeName].size,
+            maxLength: ATTR.size,
             ...baseAttrs
           }
         })
@@ -236,7 +243,7 @@ export default {
       {
         attrs: {
           label: this.label,
-          required: this.isRequired,
+          required: defIsRequired,
           error: this.isError,
           ...this.$attrs
         }